Moved the ini folder into the newly created dflowfm folder and updated the references.
