+++ hydrolib/core/io/obs/models.py
@@ -3,8 +3,8 @@ from typing import List, Literal, Optional
 from pydantic.fields import Field
 
 from hydrolib.core.io.dflowfm.common.models import LocationType
-from hydrolib.core.io.ini.models import INIBasedModel, INIGeneral, INIModel
-from hydrolib.core.io.ini.util import (
+from hydrolib.core.io.dflowfm.ini.models import INIBasedModel, INIGeneral, INIModel
+from hydrolib.core.io.dflowfm.ini.util import (
     LocationValidationConfiguration,
     LocationValidationFieldNames,
     get_enum_validator,
