+++ hydrolib/core/io/obscrosssection/models.py
@@ -2,8 +2,8 @@ from typing import List, Literal, Optional
 
 from pydantic.fields import Field
 
-from hydrolib.core.io.ini.models import INIBasedModel, INIGeneral, INIModel
-from hydrolib.core.io.ini.util import (
+from hydrolib.core.io.dflowfm.ini.models import INIBasedModel, INIGeneral, INIModel
+from hydrolib.core.io.dflowfm.ini.util import (
     LocationValidationConfiguration,
     get_location_specification_rootvalidator,
     get_split_string_on_delimiter_validator,
