+++ tests/io/test_ext.py
@@ -6,7 +6,7 @@ from pydantic import ValidationError
 
 from hydrolib.core.io.dflowfm.bc.models import Constant, ForcingModel, RealTime
 from hydrolib.core.io.dflowfm.ext.models import Boundary, ExtModel, Lateral
-from hydrolib.core.io.ini.models import INIBasedModel
+from hydrolib.core.io.dflowfm.ini.models import INIBasedModel
 
 from ..utils import test_data_dir
