+++ tests/io/test_structure.py
@@ -7,7 +7,7 @@ import pytest
 from pydantic.error_wrappers import ValidationError
 
 from hydrolib.core.io.friction.models import FrictionType
-from hydrolib.core.io.ini.parser import Parser, ParserConfig
+from hydrolib.core.io.dflowfm.ini.parser import Parser, ParserConfig
 from hydrolib.core.io.structure.models import (
     Bridge,
     Compound,
